♻️ refactor: seperate SearchBar
